requirements.txt
streamlit
ultralytics
pandas
numpy
Pillow
requests
